refactor(fl): extract asset pipeline into fl/asset/ module
The v1 asset pipeline (FL_ASSET, fl::asset_detail, manifest resolution
for WASM/host) is a self-contained subsystem and now gets its own
module slot at fl/asset/.

- Moves fl/asset.h + fl/asset.cpp.hpp -> fl/asset/
- Adds fl/asset/_build.cpp.hpp (module unity-build aggregator)
- Adds src/fl/build/fl.asset+.cpp (per-module recursive aggregator)
- Removes flat asset includes from src/fl/_build.cpp.hpp and
  src/fl/build/fl.cpp; replaces with sub-directory aggregator pull
- Sweeps consumer includes from "fl/asset.h" to "fl/asset/asset.h"
