+++ src/fl/asset/_build.cpp.hpp
@@ -1,0 +1,4 @@
+/// @file _build.cpp.hpp
+/// @brief Unity build header for fl/asset/ directory
+
+#include "fl/asset/asset.cpp.hpp"
