+++ src/fl/asset/asset.cpp.hpp
@@ -10,7 +10,7 @@
 /// fallback branch below, which currently returns an empty `fl::url`.
 /// TODO(#2284): ESP32 LittleFS / SD-card resolution is future work.
 
-#include "fl/asset.h"
+#include "fl/asset/asset.h"
 #include "fl/stl/fstream.h"
 #include "fl/stl/noexcept.h"
 #include "fl/stl/string.h"
